Extract remaining model page sections
